feat(hub-common): add itemToContent and related functions
affects: @esri/hub-common, @esri/hub-content
