+++ packages/content/test/mocks/items
@@ -1,0 +1,1 @@
+../../../common/test/mocks/items
