+++ packages/content/test/portal.test.ts
@@ -1,19 +1,10 @@
 import * as fetchMock from "fetch-mock";
 import * as arcgisRestPortal from "@esri/arcgis-rest-portal";
 import { IItem } from "@esri/arcgis-rest-portal";
-import { IEnvelope } from "@esri/arcgis-rest-types";
-import { IHubRequestOptions, cloneObject, IHubContent } from "@esri/hub-common";
-import * as commonModule from "@esri/hub-common";
-import {
-  getContentFromPortal,
-  itemToContent,
-  parseItemCategories,
-  getItemHubType,
-  getFamily,
-} from "../src/portal";
+import { IHubRequestOptions, IHubContent } from "@esri/hub-common";
+import { getContentFromPortal } from "../src/portal";
 import * as metadataModule from "../src/metadata";
 import * as documentItem from "./mocks/items/document.json";
-import * as mapServiceItem from "./mocks/items/map-service.json";
 import { mockUserSession } from "./test-helpers/fake-user-session";
 
 function validateContentFromPortal(content: IHubContent, item: IItem) {
@@ -63,110 +54,6 @@ function validateContentFromPortal(content: IHubContent, item: IItem) {
   expect(content.isDownloadable).toBe(true);
 }
 
-describe("item to content", () => {
-  let item: IItem;
-  beforeEach(() => {
-    item = cloneObject(documentItem) as IItem;
-  });
-  it("gets summary from description when no snippet", () => {
-    item.snippet = null;
-    const content = itemToContent(item);
-    expect(content.summary).toBe(item.description);
-  });
-  it("gets permissions.control from itemControl when it exists", () => {
-    item.itemControl = "update";
-    const content = itemToContent(item);
-    expect(content.permissions.control).toBe(item.itemControl);
-  });
-  it("attaches normalizedItemType", () => {
-    spyOn(commonModule, "normalizeItemType").and.returnValue("Normalized Type");
-    const content = itemToContent(item);
-    expect(content.normalizedType).toBe("Normalized Type");
-  });
-  describe("when item has properties", () => {
-    it("should set actionLinks to links", () => {
-      item.properties = {
-        links: [{ url: "https://foo.com" }],
-      };
-      const content = itemToContent(item);
-      expect(content.actionLinks).toEqual(item.properties.links);
-    });
-  });
-  it("has a reference to the item", () => {
-    const content = itemToContent(item);
-    expect(content.item).toBe(item);
-  });
-  it("has a boundary when the item has a valid extent", () => {
-    item = cloneObject(mapServiceItem) as IItem;
-    const content = itemToContent(item);
-    const geometry: IEnvelope = {
-      xmin: -2.732,
-      ymin: 53.4452,
-      xmax: -2.4139,
-      ymax: 53.6093,
-      spatialReference: {
-        wkid: 4326,
-      },
-    };
-    expect(content.boundary).toEqual({ geometry });
-  });
-  // NOTE: other use cases (including when a portal is passed)
-  // are covered by getContentFromPortal() tests
-});
-describe("get item family", () => {
-  it("returns dataset for image service", () => {
-    expect(getFamily("Image Service")).toBe("dataset");
-  });
-  it("returns map for feature service and raster layer", () => {
-    expect(getFamily("Feature Service")).toBe("map");
-    expect(getFamily("Raster Layer")).toBe("map");
-  });
-  it("returns document for excel", () => {
-    expect(getFamily("Microsoft Excel")).toBe("document");
-  });
-  it("returns template for solution", () => {
-    expect(getFamily("Solution")).toBe("template");
-  });
-  it("returns content for other specific types", () => {
-    expect(getFamily("CAD Drawing")).toBe("content");
-    expect(getFamily("Feature Collection Template")).toBe("content");
-    expect(getFamily("Report Template")).toBe("content");
-  });
-  it("returns content for collection other", () => {
-    expect(getFamily("360 VR Experience")).toBe("content");
-  });
-});
-describe("get item hub type", () => {
-  it("normalizes item", () => {
-    expect(
-      getItemHubType({
-        type: "Hub Initiative",
-        typeKeywords: ["hubInitiativeTemplate"],
-      } as IItem)
-    ).toBe("template");
-  });
-  it("works with just type", () => {
-    expect(getItemHubType("Form")).toBe("feedback");
-  });
-});
-describe("parse item categories", () => {
-  it("parses the categories", () => {
-    const categories = [
-      "/Categories/Boundaries",
-      "/Categories/Planning and cadastre/Property records",
-      "/Categories/Structure",
-    ];
-    expect(parseItemCategories(categories)).toEqual([
-      "Boundaries",
-      "Planning and cadastre",
-      "Property records",
-      "Structure",
-    ]);
-  });
-  it("doesn't blow up with undefined", () => {
-    expect(() => parseItemCategories(undefined)).not.toThrow();
-  });
-});
 describe("get content from portal", () => {
   // emulate the get groups response
   const mockItemGroups: any = {
